KNN Classifier Example
Source:: https://towardsdatascience.com/k-nearest-neighbor-python-2fccc47d2a55

# Problem Statement:: 
Classifying data using the K-Nearest Neighbors algorithm in Python. For this tutorial, we’ll be using the breast cancer dataset from the sklearn.datasets module.
        
The dataset classifies tumors into two categories (malignant and benign) and contains something like 30 features. 
In the real world, you’d look at the correlations and select a subset of features that plays the greatest role in 
determining whether a tumor is malignant or not. However, for the sake of simplicity, we’ll pick a couple at random. 
We must encode categorical data for it to be interpreted by the model (i.e. malignant = 0 and benign = 1).