KNN Regressor
Source: https://realpython.com/knn-python/

Problem Statement:: Use kNN to Predict the Age of Sea Slugs

To follow along with the coding part, you’ll look at an example dataset for the rest of this tutorial—the Abalone Dataset. This dataset contains age measurements on a large number of abalones.